% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{node10_norm}
\alias{node10_norm}
\title{Normalized adjacency matrix using node10 object}
\format{
Matrix sparseMatrix
}
\usage{
node10_norm
}
\description{
A column-wise normalized adjacency sparse matrix
}
\keyword{datasets}
