% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{node10_rwr}
\alias{node10_rwr}
\title{Example results from RWR}
\format{
a list include whether RWR converges within given iterations, and a vector for the post-scores for each node
}
\usage{
node10_rwr
}
\description{
Results generated from RWR with example graph and random initial scores
}
\keyword{datasets}
