% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{norm_W}
\alias{norm_W}
\title{Normalize adjacency matrix from igraph graph object}
\usage{
norm_W(graph)
}
\arguments{
\item{graph}{igraph graph object}
}
\value{
A column-wise normalized matrix
}
\description{
Normalize adjacency matrix from igraph graph object
}
\examples{
node10
norm_W(node10)
}
