% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{runRWR_normAdj}
\alias{runRWR_normAdj}
\title{Run random walk with restart (RWR)}
\usage{
runRWR_normAdj(
  values,
  adj_mat,
  int_scores = NULL,
  restart_prop = 0.15,
  num_iter = 1000,
  delta = 1e-06
)
}
\arguments{
\item{values}{a vector for initial scores, the orders should match node orders}

\item{adj_mat}{column-wise normalized adjacency (sparse) matrix}

\item{int_scores}{NULL or vector, intermediate scores if run the function on the fly}

\item{restart_prop}{0.15, restart probability}

\item{num_iter}{maximum iterations to process}

\item{delta}{minimal differences for convergence}
}
\value{
a list include (1) whether RWR converges in defined iterations; (2) output vector after RWR
}
\description{
Run random walk with restart (RWR)
}
\examples{
scores = runif(ncol(node10_norm))
runRWR_normAdj(values = scores, adj_mat = node10_norm)
}
