% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stringdb_human}
\alias{stringdb_human}
\title{Largest connected community from human String database}
\format{
igraph object
}
\source{
\url{https://stringdb-downloads.org/}
}
\usage{
stringdb_human
}
\description{
A igraph object for the largest connected community derived from human String database. The connections were pruned to only retain
confidence scores greater than 900 (or 0.9).
}
\keyword{datasets}
