% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{stringdb_human_norm}
\alias{stringdb_human_norm}
\title{Normalized adjacency matrix for the largest connected community from human String database}
\format{
sparse matrix
}
\source{
\url{https://stringdb-downloads.org/}
}
\usage{
stringdb_human_norm
}
\description{
A sparse matrix for the largest connected community derived from human String database. The connections were pruned to only retain
confidence scores greater than 900 (or 0.9). The connections were subsequently column-wisely normalized and converted into sparse matrix.
}
\keyword{datasets}
