% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{testInput}
\alias{testInput}
\title{Example data for RWR input}
\format{
data frame: gene_name score
}
\usage{
testInput
}
\description{
A dataframe to test the RWR function. In practice, only use the second column in the function. Genes are ordered to match the graph object.
}
\keyword{datasets}
